Change function VariablesNoIteration
The code has an error when calling function "solveIteration", the issue was present due to the output of VariablesNoIteration that are numbers with a dot at the end (float numbers). Change on the function to create np.array of integers instead.
Root cause: failure in for i in VariablesNoIteration  due to the float numbers

+++ UpSimplex/Simplex.py
@@ -507,7 +507,7 @@ def variablesNoiteration(matrix, variablesIteration):
     # set of all variables
     groupOfVariables = np.array(range(matrix.shape[1]))
     # variables which are not in the iteration
-    varNoIter = np.array([])
+    varNoIter = np.array([], dtype='int16')
 
     # It is checked which variables are not in the iteration
     for i in groupOfVariables:
